Added readWithTimeout as a possible fix to #34
This makes the callback coroutine cancellable and then adds a timeout to it. It makes sure the mutex is released on cancellation and also checks if it is connected before actually trying to read.
This is not ready to be merged but it should serve as a discussion on how to fix #34